Refactor mudata tests
