+++ tests/models/test_models.py
@@ -8,7 +8,6 @@ import torch
 
 import scvi
 from scvi.data import synthetic_iid
-from scvi.data._compat import LEGACY_REGISTRY_KEY_MAP
 from scvi.data._download import _download
 from scvi.dataloaders import (
     AnnDataLoader,
@@ -28,51 +27,6 @@ from scvi.model import (
 from scvi.train import TrainingPlan, TrainRunner
 from scvi.utils import attrdict
 from tests.dataset.utils import generic_setup_adata_manager, scanvi_setup_adata_manager
-
-LEGACY_REGISTRY_KEYS = set(LEGACY_REGISTRY_KEY_MAP.values())
-LEGACY_SETUP_DICT = {
-    "scvi_version": "0.0.0",
-    "categorical_mappings": {
-        "_scvi_batch": {
-            "original_key": "testbatch",
-            "mapping": np.array(["batch_0", "batch_1"], dtype=object),
-        },
-        "_scvi_labels": {
-            "original_key": "testlabels",
-            "mapping": np.array(["label_0", "label_1", "label_2"], dtype=object),
-        },
-    },
-    "extra_categoricals": {
-        "mappings": {
-            "cat1": np.array([0, 1, 2, 3, 4]),
-            "cat2": np.array([0, 1, 2, 3, 4]),
-        },
-        "keys": ["cat1", "cat2"],
-        "n_cats_per_key": [5, 5],
-    },
-    "extra_continuous_keys": np.array(["cont1", "cont2"], dtype=object),
-    "data_registry": {
-        "X": {"attr_name": "X", "attr_key": None},
-        "batch_indices": {"attr_name": "obs", "attr_key": "_scvi_batch"},
-        "labels": {"attr_name": "obs", "attr_key": "_scvi_labels"},
-        "cat_covs": {
-            "attr_name": "obsm",
-            "attr_key": "_scvi_extra_categoricals",
-        },
-        "cont_covs": {
-            "attr_name": "obsm",
-            "attr_key": "_scvi_extra_continuous",
-        },
-    },
-    "summary_stats": {
-        "n_batch": 2,
-        "n_cells": 400,
-        "n_vars": 100,
-        "n_labels": 3,
-        "n_proteins": 0,
-        "n_continuous_covs": 2,
-    },
-}
 
 
 def test_saving_and_loading(save_path):
